[core] state functions return (Commit, Key), delete module
Nearly every callsite cloned the Keys as it is.

Divides its contents between the other areas.
