+++ dom/examples/hacking/src/lib.rs
@@ -22,13 +22,13 @@ pub fn begin() {
 
 #[topo::nested]
 fn root() -> Div {
-    let count = state(|| 0);
+    let (count, set_count) = state(|| 0);
 
     let mut root = div();
 
     root = root.child(mox! { <div>{% "hello world from moxie! ({})", &count }</div> });
     root = root.child(mox! {
-        <button type="button" onclick={move |_| count.update(|c| Some(c + 1))}>
+        <button type="button" onclick={move |_| set_count.update(|c| Some(c + 1))}>
             "increment"
         </button>
     });
